Collection Menu Affordance (BL-10943)
