+++ src/BloomBrowserUI/collectionsTab/BooksOfCollection.tsx
@@ -5,21 +5,10 @@ import Grid from "@material-ui/core/Grid";
 import React = require("react");
 import "BooksOfCollection.less";
 import { BloomApi } from "../utils/bloomApi";
-import Menu from "@material-ui/core/Menu";
-import MenuItem from "@material-ui/core/MenuItem";
-import NestedMenuItem from "material-ui-nested-menu-item";
 import { BookButton, bookButtonHeight, bookButtonWidth } from "./BookButton";
-import { useMonitorBookSelection } from "../app/selectedBook";
-import { element } from "prop-types";
-import { useL10n } from "../react_components/l10nHooks";
-import { useEffect, useState } from "react";
-import { useSubscribeToWebSocketForEvent } from "../utils/WebSocketManager";
-import { Divider, ListItemIcon, ListItemText } from "@material-ui/core";
-import { BookSelectionManager, useIsSelected } from "./bookSelectionManager";
-import LazyLoad, { forceCheck } from "react-lazyload";
+import { BookSelectionManager } from "./bookSelectionManager";
+import LazyLoad from "react-lazyload";
 import { Link } from "../react_components/link";
-import { ApiCheckbox } from "../react_components/ApiCheckbox";
-import { ApiBackedCheckbox } from "../react_components/apiBackedCheckbox";
 
 export interface IBookInfo {
     id: string;
@@ -74,109 +63,6 @@ export const BooksOfCollection: React.FunctionComponent<{
     collection.isEditableCollection = props.isEditableCollection;
     collection.id = props.collectionId;
 
-    const [contextMousePoint, setContextMousePoint] = React.useState<
-        | {
-              mouseX: number;
-              mouseY: number;
-          }
-        | undefined
-    >();
-
-    const setAdjustedContextMenuPoint = (x: number, y: number) => {
-        setContextMousePoint({
-            mouseX: x - 2,
-            mouseY: y - 4
-        });
-    };
-
-    const handleContextMenuClick = (
-        event: React.MouseEvent<HTMLDivElement>
-    ) => {
-        // Clicks on a book button should not open this menu. Normally the book button
-        // intercepts them itself, but we've had this problem when another book button
-        // is open. I think the following check is probably enough by itself, but this
-        // extra one can't hurt.
-        let badTargetParent = (event.target as HTMLElement).closest(
-            ".book-button"
-        );
-        if (badTargetParent) {
-            return;
-        }
-
-        // If there's already a Material-UI menu open somewhere (probably one of the book buttons),
-        // we don't want to open our own menu, and it's tricky to make it open the menu appropriate
-        // to the place clicked, which might not even be what the user wanted. Instead, just
-        // simulate a regular click on the backdrop, which will close the menu. If necessary the
-        // user can then right-click again.
-        // This is a known bug in Material-UI, fixed in version 5, so we may be able to do better
-        // when we switch to that. See https://github.com/mui/material-ui/issues/19145.
-        var menuBackdrop = (event.target as HTMLElement).closest(
-            ".MuiPopover-root"
-        );
-        if (menuBackdrop) {
-            (event.target as HTMLElement).click();
-            return;
-        }
-        if (props.isEditableCollection) {
-            setAdjustedContextMenuPoint(event.clientX - 2, event.clientY - 4);
-            event.preventDefault();
-            event.stopPropagation();
-        }
-    };
-
-    const handleClose = () => {
-        setContextMousePoint(undefined);
-    };
-
-    const collectionMenuItemsSpecs: MenuItemSpec[] = [
-        {
-            label: "Open or Create Another Collection",
-            l10nId: "CollectionTab.OpenCreateCollectionMenuItem",
-            command: "workspace/openOrCreateCollection"
-        },
-        {
-            label: "Make Reader Template Bloom Pack...",
-            l10nId:
-                "CollectionTab.AddMakeReaderTemplateBloomPackToolStripMenuItem",
-            command: "collections/makeBloompack"
-        },
-        {
-            label: "Advanced",
-            l10nId: "CollectionTab.AdvancedToolStripMenuItem",
-            shouldShow: () => true, // show for all collections (except factory)
-            submenu: [
-                {
-                    label: "Do Checks of All Books",
-                    l10nId: "CollectionTab.CollectionMenu.doChecksOfAllBooks",
-                    command: "collections/doChecksOfAllBooks"
-                },
-                {
-                    label: "Rescue Missing Images...",
-                    l10nId: "CollectionTab.CollectionMenu.rescueMissingImages",
-                    command: "collections/rescueMissingImages"
-                },
-                {
-                    label: "Do Updates of All Books",
-                    l10nId:
-                        "CollectionTab.CollectionMenu.doChecksAndUpdatesOfAllBooks",
-                    command: "collections/doUpdatesOfAllBooks"
-                }
-            ]
-        }
-    ];
-
-    //const bookMenuItems = makeMenuItems(bookMenuItemsSpecs);
-    const collectionMenuItems = makeMenuItems(
-        collectionMenuItemsSpecs,
-        props.isEditableCollection,
-        props.manager.getSelectedBookInfo()?.saveable ?? false,
-        handleClose,
-        // the collection menu commands don't actually use the ID of
-        // a particular book
-        "",
-        props.collectionId
-    );
-
     // This is an approximation. 5 buttons per line is about what we get in a default
     // layout on a fairly typical screen. We'd get a better approximation if we used
     // the width of a button and knew the width of the container. But I think this is good
@@ -187,7 +73,6 @@ export const BooksOfCollection: React.FunctionComponent<{
         <div
             key={"BookCollection-" + props.collectionId}
             className="bookButtonPane"
-            onContextMenu={e => handleContextMenuClick(e)}
             style={{ cursor: "context-menu" }}
         >
             {books.length > 0 && (
@@ -255,20 +140,6 @@ export const BooksOfCollection: React.FunctionComponent<{
                     Get more source books at BloomLibrary.org
                 </Link>
             )}
-            {contextMousePoint && (
-                <Menu
-                    keepMounted={true}
-                    open={contextMousePoint !== undefined}
-                    onClose={handleClose}
-                    anchorReference="anchorPosition"
-                    anchorPosition={{
-                        top: contextMousePoint!.mouseY,
-                        left: contextMousePoint!.mouseX
-                    }}
-                >
-                    {collectionMenuItems}
-                </Menu>
-            )}
         </div>
     );
     // There's no point in lazily loading an empty list of books. But more importantly, on early renders
@@ -287,194 +158,5 @@ export const BooksOfCollection: React.FunctionComponent<{
         </LazyLoad>
     ) : (
         content
-    );
-};
-
-export interface MenuItemSpec {
-    label: string;
-    l10nId?: string;
-    // One of these two must be provided. If both are, onClick is used and command is ignored.
-    // If only command is provided, the click action is to call handleBookCommand with that argument,
-    // which invokes the corresponding API call to C# code.
-    command?: string;
-    onClick?: React.MouseEventHandler<HTMLElement>;
-    // If this is defined (rare), it determines whether the menu item should be shown
-    // (except in factory collections, where we never show any).
-    // If it's not defined, a menu item is shown if we're in the editable collection and
-    // other requirements are satisfied, and not otherwise.
-    shouldShow?: () => boolean;
-    // Involves making changes to the book; therefore, can only be done in the one editable collection
-    // (unless shouldInclude returns true), and if we're in a Team Collection, the book must be checked out.
-    requiresSavePermission?: boolean;
-    submenu?: MenuItemSpec[];
-    icon?: React.ReactNode;
-    // if true, menu item is rendered as an ApiCheckbox with the command as its api.
-    checkbox?: boolean;
-}
-
-// This function and the associated MenuItem classes want to become a general component for making
-// pop-up menus. But at the moment a lot of the logic is specific to making menus about books and
-// book collections. I'm not seeing a good way to factor that out. Maybe it will become clear when
-// we have a third need for such a menu. For now it is just logic shared with BookButton.
-export const makeMenuItems = (
-    menuItemsSpecs: MenuItemSpec[],
-    isEditableCollection: boolean,
-    isBookSavable: boolean,
-    close: () => void,
-    bookId: string,
-    collectionId: string
-) => {
-    const menuItemsT = menuItemsSpecs
-        .map((spec: MenuItemSpec) => {
-            if (spec.label === "-") {
-                return <Divider />;
-            }
-            if (spec.submenu) {
-                var submenuItems = makeMenuItems(
-                    spec.submenu,
-                    isEditableCollection,
-                    isBookSavable,
-                    close,
-                    bookId,
-                    collectionId
-                );
-                return (
-                    <LocalizableNestedMenuItem
-                        english={spec.label}
-                        l10nId={spec.l10nId!}
-                    >
-                        {submenuItems}
-                    </LocalizableNestedMenuItem>
-                );
-            }
-            if (spec.shouldShow) {
-                if (!spec.shouldShow()) {
-                    return undefined;
-                }
-            } else {
-                // default logic for whether to show the command
-                if (isEditableCollection) {
-                    // eliminate commands that require permission to change the book, if we don't have it
-                    if (spec.requiresSavePermission && !isBookSavable) {
-                        return undefined;
-                    }
-                } else {
-                    // outside that collection, commands can only be shown if they have a shouldShow function.
-                    return undefined;
-                }
-            }
-
-            if (spec.checkbox) {
-                return (
-                    <LocalizableCheckboxMenuItem
-                        english={spec.label}
-                        l10nId={spec.l10nId!}
-                        onClick={() => close()}
-                        apiEndpoint={spec.command!}
-                        icon={spec.icon}
-                    ></LocalizableCheckboxMenuItem>
-                );
-            }
-            // It should be possible to use spec.onClick || () => handleBookCommand(spec.command!) inline,
-            // but I can't make Typescript accept it.
-            let clickAction: React.MouseEventHandler = () => {
-                close();
-                BloomApi.postString(
-                    `${spec.command!}?collection-id=${encodeURIComponent(
-                        collectionId
-                    )}`,
-                    bookId
-                );
-            };
-            if (spec.onClick) {
-                clickAction = spec.onClick;
-            }
-            return (
-                <LocalizableMenuItem
-                    english={spec.label}
-                    l10nId={spec.l10nId!}
-                    onClick={clickAction}
-                    icon={spec.icon}
-                ></LocalizableMenuItem>
-            );
-        })
-        .filter(x => x); // that is, remove ones where the map function returned undefined
-
-    // Can't find a really good way to tell that an element is a Divider.
-    // But we only have Dividers and LocalizableMenuItems in this list,
-    // so it's a Dividier if it doesn't have one of the required props of LocalizableMenuItem.
-    const isDivider = (element: JSX.Element): boolean => {
-        return !element.props.english;
-    };
-    // filter out dividers if (a) followed by another divider, or (b) at the start or end of the list
-    return menuItemsT.filter(
-        (elt, index) =>
-            !isDivider(elt!) ||
-            (index > 0 &&
-                index < menuItemsT.length - 1 &&
-                !isDivider(menuItemsT[index + 1]!))
-    );
-};
-
-const LocalizableMenuItem: React.FunctionComponent<{
-    english: string;
-    l10nId: string;
-    onClick: React.MouseEventHandler<HTMLElement>;
-    icon?: React.ReactNode;
-}> = props => {
-    const label = useL10n(props.english, props.l10nId);
-    return (
-        <MenuItem key={props.l10nId} onClick={props.onClick}>
-            {props.icon ? (
-                <React.Fragment>
-                    <ListItemIcon
-                        css={css`
-                            min-width: 30px !important; // overrides MUI default that leaves way too much space
-                        `}
-                    >
-                        {props.icon}
-                    </ListItemIcon>
-                    <ListItemText>{label}</ListItemText>
-                </React.Fragment>
-            ) : (
-                label
-            )}
-        </MenuItem>
-    );
-};
-
-const LocalizableCheckboxMenuItem: React.FunctionComponent<{
-    english: string;
-    l10nId: string;
-    onClick: React.MouseEventHandler<HTMLElement>;
-    apiEndpoint: string;
-    icon?: React.ReactNode;
-}> = props => {
-    const label = useL10n(props.english, props.l10nId);
-    return (
-        <MenuItem key={props.l10nId} onClick={props.onClick}>
-            <ApiBackedCheckbox
-                l10nKey={props.l10nId!}
-                apiEndpoint={props.apiEndpoint}
-            >
-                {label}
-            </ApiBackedCheckbox>
-        </MenuItem>
-    );
-};
-
-const LocalizableNestedMenuItem: React.FunctionComponent<{
-    english: string;
-    l10nId: string;
-}> = props => {
-    const label = useL10n(props.english, props.l10nId);
-    return (
-        // Can't find any doc on parentMenuOpen. Examples set it to the same value
-        // as the open prop of the parent menu. But it seems to work fine just set
-        // to true. (If omitted, however, the child menu does not appear when the
-        // parent is hovered over.)
-        <NestedMenuItem key={props.l10nId} label={label} parentMenuOpen={true}>
-            {props.children}
-        </NestedMenuItem>
     );
 };
