+++ src/BloomBrowserUI/react_components/icons/ImportIcon.tsx
@@ -1,7 +1,7 @@
 import React = require("react");
-import SvgIcon from "@material-ui/core/SvgIcon";
+import SvgIcon, { SvgIconProps } from "@material-ui/core/SvgIcon";
 
-export const ImportIcon = props => (
+export const ImportIcon = (props: SvgIconProps) => (
     <SvgIcon {...props}>
         <path
             d="m4.8 1e-6c-1.332 0-2.4 1.08-2.4 2.4v19.2c0 0.63648 0.25286 1.2469 0.70294 1.697 0.45009 0.45012 1.0605 0.70296 1.6971 0.70296h14.4c0.63648 0 1.2469-0.25284 1.697-0.70296s0.70296-1.0606 0.70296-1.697v-14.4l-7.2-7.2zm8.4 1.8 6.6 6.6h-6.6zm-3.5763 12.129 3.3891-3.3844 3.4056 3.3766-3.3799 3.3799 2.5125 2.5628h-8.4725v-8.4709"
